fix: trust api happy impl
